requirements.txt
anyio==3.6.2
autopep8==2.0.2
click==8.1.3
colorama==0.4.6
fastapi==0.95.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
pycodestyle==2.10.0
pydantic==1.10.7
PyJWT==2.6.0
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.21.1
watchfiles==0.19.0
websockets==11.0.1
